build send to client

+++ app/src/main/java/com/app/abcdapp/activities/WithdrawalActivity.java
@@ -84,10 +84,10 @@ public class WithdrawalActivity extends AppCompatActivity {
                 if (etAmount.getText().toString().trim().equals("") || etAmount.getText().toString().trim().equals("0")){
                     etAmount.setError("enter amount");
                     etAmount.requestFocus();
-                }else if (Integer.parseInt(etAmount.getText().toString().trim()) < 250) {
+                }else if (Double.parseDouble(etAmount.getText().toString().trim()) < 250) {
                     Toast.makeText(activity, "minimum 250 balance required", Toast.LENGTH_SHORT).show();
 
-                }else if (Integer.parseInt(etAmount.getText().toString().trim()) > Integer.parseInt(session.getData(Constant.BALANCE))) {
+                }else if (Double.parseDouble(etAmount.getText().toString().trim()) > Double.parseDouble(session.getData(Constant.BALANCE))) {
                     Toast.makeText(activity, "insuffcient balance", Toast.LENGTH_SHORT).show();
 
                 }else {
